requirements.txt
asgiref==3.3.4
astroid==2.5.3
dj-database-url==0.5.0
Django==3.0.1
django-heroku==0.3.1
flake8==3.9.1
gunicorn==19.9.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.7.4
python-dotenv==0.17.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==4.1.4
wrapt==1.12.1
